add alsa pcm backend for mrg
The backend accepts both floating point and signed 16 bit stereo and mono
clients at arbitrary sample-rates, with a fixed output to 48000 stereo. This
makes the gtk version of mrg which doesn't use mmm have a realtime audio API to
use which is equivalent and controlled by the same entry points.