make mrg_queue_draw() work during painting
A separate dirty during painting rectangle is maintained and at the
end of painting instead of clearing the dirty rectangle the dirty
rectangle gets the contents of the dirty_during_paint rectangle and
the dirty_during_paint rectangle is cleared.

This permits animations to queue a re-render at next possible frame, this will
yield better results than hacks using idle callbacks to queue a subsequent
render after flush is done.